+++ node/clients/checkout.ts
@@ -5,6 +5,7 @@ import {
   JanusClient,
   RequestConfig,
 } from '@vtex/api'
+import { UserProfileInput } from 'vtex.checkout-graphql'
 
 import { checkoutCookieFormat, statusToError } from '../utils'
 
@@ -86,21 +87,24 @@ export class Checkout extends JanusClient {
     )
 
   public updateOrderFormPayment = (orderFormId: string, payments: any) =>
-    this.post<OrderForm>(
+    this.post<CheckoutOrderForm>(
       this.routes.attachmentsData(orderFormId, 'paymentData'),
       { payments },
       { metric: 'checkout-updateOrderFormPayment' }
     )
 
-  public updateOrderFormProfile = (orderFormId: string, fields: any) =>
-    this.post<OrderForm>(
+  public updateOrderFormProfile = (
+    orderFormId: string,
+    fields: UserProfileInput
+  ) =>
+    this.post<CheckoutOrderForm>(
       this.routes.attachmentsData(orderFormId, 'clientProfileData'),
       fields,
       { metric: 'checkout-updateOrderFormProfile' }
     )
 
   public updateOrderFormShipping = (orderFormId: string, shipping: any) =>
-    this.post<OrderForm>(
+    this.post<CheckoutOrderForm>(
       this.routes.attachmentsData(orderFormId, 'shippingData'),
       shipping,
       { metric: 'checkout-updateOrderFormShipping' }
@@ -110,7 +114,7 @@ export class Checkout extends JanusClient {
     orderFormId: string,
     marketingData: any
   ) =>
-    this.post<OrderForm>(
+    this.post<CheckoutOrderForm>(
       this.routes.attachmentsData(orderFormId, 'marketingData'),
       marketingData,
       { metric: 'checkout-updateOrderFormMarketingData' }