+++ node/index.ts
@@ -434,6 +434,20 @@ declare global {
   interface SavePaymentTokenPayload {
     status: string
   }
+
+  interface UserProfileInput {
+    email?: string
+    firstName?: string
+    lastName?: string
+    document?: string
+    phone?: string
+    documentType?: string
+    isCorporate?: boolean
+    corporateName?: string
+    tradeName?: string
+    corporateDocument?: string
+    stateInscription?: string
+  }
 }
 
 // Segments are small and immutable.